VH-68: four faults that were nobody's ticket
The review's consolidation dropped these, and each survives precisely
because it is too small to schedule.

`SlidingMinimum.position` counts samples for the length of a file and
never resets, in an `Int32Array` that wraps past 2^31 — 12.4 hours at
48 kHz. After the wrap the expiry comparison goes negative and the ring
cycles forever. Well outside this tool's envelope; a latent hang is still
a latent hang. `Float64Array` holds every integer to 2^53 exactly, which
is 285,000 years of audio. Pinned by pushing the counter past the old
boundary directly, because clocking 2.1 billion samples through is not a
test.

`WARNING_THRESHOLDS.clippingDbtp` and `COMPRESSOR.softKnee` were declared
and never read, while `truepeak.ts` carried its own `-0.1` and
`compressor.ts` its own `6`. That is worse than a plain literal: a
literal admits where the number lives, while a config entry nobody reads
invites someone to tune it and watch nothing happen. `softKnee: true`
also described the shape while another file decided the width, so it
becomes `kneeDb: 6` — a soft knee is how wide it is. Both are now pinned
by reading the behaviour back, not the constant.

`extended-silence` sat inside a guard written for the NOISE test. That
guard exists so a recording with no pauses is not accused of background
noise it may not have — a judgement about noise, applied by accident to
silence. An entirely silent track has every short-term value at
`-Infinity`, so the guard emptied and the one warning that describes it
could never fire. It moves out; the noise test keeps its guard.

And `run-in-engines.mjs` printed `wanted.length - failures` complete
runs, so "3/3" could mean one ran and two were not installed. Three
independent counters now, and a skip fails the run only when that engine
was named explicitly — defaulting to all three means "whatever is
installed"; naming one means "this one".

Protected DSP touched (`truepeak.ts`, `limiter.ts`), so the EBU Tech 3341
harness was re-run in the same change: 22 passed, 1 skipped, unchanged.
npm run check: 413 passed, 1 skipped, 0 broken links.

+++ src/audio/limiter.ts
@@ -28,14 +28,22 @@ const MINIMUM_MAGNITUDE = 1e-12
  */
 class SlidingMinimum {
   private readonly values: Float64Array
-  private readonly indices: Int32Array
+  /**
+   * Float64, not Int32. `position` counts samples for the length of the file
+   * and never resets, so an `Int32Array` wraps past 2^31 — about 12.4 hours at
+   * 48 kHz — after which `indices[head] < oldest` compares a negative number
+   * and the expiry loop cycles the whole ring forever. Outside the envelope
+   * this tool is built for, and a latent hang is still a latent hang (VH-68).
+   * A double holds every integer to 2^53 exactly: 285,000 years of audio.
+   */
+  private readonly indices: Float64Array
   private head = 0
   private tail = 0
   private position = 0
 
   constructor(private readonly windowSize: number) {
     this.values = new Float64Array(windowSize + 1)
-    this.indices = new Int32Array(windowSize + 1)
+    this.indices = new Float64Array(windowSize + 1)
   }
 
   push(value: number): number {
